Fix import in client + move util test to nodejs
